requirements.txt
bs4
yake
spacy
flask
ollama
chromadb
langchain
langchain_community
googlesearch-python